Close #10: Folder icon looks terrible on retina devices
We could solve this issue by adding appropriate `srcset` attributes as
for the thumbnails, but using an SVG image is much simpler, especially
for the webdesigner. Of course, SVG images are not supported by old
browsers (IE 8, for instance), and possible
[workarounds](https://css-tricks.com/a-complete-guide-to-svg-fallbacks/)
are unrewarding, so we leave our users holding the baby, i.e. we offer
a PNG and an SVG icon, and let the user decide which one to use.