Close #36: Cater to multilingual sites
Instead of using embedded metadata for the image captions, we read
these from a file `foldergallery.php` if it is available in the
respective folder.  The file has to return an array of language codes
mapping to an array of filenames mapping to the caption.  For instance:

````php
return [
    'en' => [
        'pic1.jpg' => 'Caption1',
        'subfolder' => 'Caption2'
    ],
    'de' => [
        'pic1.jpg' => 'Beschriftung1',
        'subfolder' => 'Beschriftung2'
];
````

This way the gallery can be used for multilingual sites, and
furthermore it is possible to have explicit captions for folders.